+++ apps/server/src/dipeo_server/api/__init__.py
@@ -3,10 +3,8 @@
 from .context import get_graphql_context
 from .middleware import setup_middleware
 from .router import setup_routes
-from .schema import create_graphql_router
 
 __all__ = [
-    "create_graphql_router",
     "get_graphql_context",
     "setup_middleware",
     "setup_routes",
