+++ apps/web/src/graphql/types/index.ts
@@ -1,25 +1,11 @@
-/**
- * GraphQL Type Exports
- * 
- * This file provides a clean interface for accessing GraphQL-generated types
- * and related utilities throughout the application.
- * 
- * CONSOLIDATION NOTE:
- * This reduces the re-declaration of GraphQL types by:
- * 1. Directly exporting generated types from @/__generated__/graphql
- * 2. Keeping semantic aliases (Domain*, React*) only in graphql-mappings for backward compatibility
- * 3. Providing a single import point for all GraphQL-related types and utilities
- * 
- * Import from here when you need GraphQL types or conversion utilities.
- */
 
 // Re-export generated GraphQL types directly
 export type {
-  DomainNode as Node,
-  DomainArrow as Arrow,
-  DomainHandle as Handle,
-  DomainPerson as Person,
-  DomainApiKey as ApiKey,
+  DomainNodeType as Node,
+  DomainArrowType as Arrow,
+  DomainHandleType as Handle,
+  DomainPersonType as Person,
+  DomainApiKeyType as ApiKey,
   DomainDiagramType,
   Vec2Input,
   // Any other types needed from GraphQL
