+++ packages/domain-models/src/diagram.ts
