+++ packages/domain-models/src/index.ts
@@ -1,4 +1,4 @@
-export * from './diagram';
-export * from './execution';
-export * from './person';
-// export * from './llm'; // Currently empty - all LLM types were unused and removed
+export * from './diagram.js';
+export * from './execution.js';
+export * from './person.js';
+export * from './conversation.js';
