--- packages/domain-models/src/llm/index.ts
@@ -1,10 +1,0 @@
-/**
- * Shared LLM domain models
- * These interfaces serve as the single source of truth for LLM-related types
- * Used by both frontend (TypeScript) and backend (Python via code generation)
- */
-
-// This module is currently empty as all LLM-related types have been moved
-// to their respective domains or removed due to being unused
-
-
